pharma_django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
reportlab==4.0.7
openpyxl==3.1.2
